+++ src/the/bytecode/club/bootloader/resource/EmptyExternalResource.java
@@ -1,9 +1,7 @@
-package the.bytecode.club.bootloader;
+package the.bytecode.club.bootloader.resource;
 
 import java.io.IOException;
 import java.net.URL;
-
-import the.bytecode.club.bootloader.resource.ExternalResource;
 
 /**
  * @author Bibl (don't ban me pls)
